@@ -1,14 +1,12 @@
 apply plugin: 'com.android.library'
-apply plugin: 'img-optimizer'
 
 dependencies {
-    compile fileTree(dir: 'libs', include: '*.jar')
+    api fileTree(dir: 'libs', include: '*.jar')
 }
 
 android {
-    compileSdkVersion 25
-    // 编译工具的版本
-    buildToolsVersion "25.0.0"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
 
     sourceSets {
         main {
@@ -21,9 +19,6 @@ android {
             assets.srcDirs = ['assets']
             jniLibs.srcDirs = ['libs']
         }
-
-        // Move the tests to tests/java, tests/res, etc...
-        instrumentTest.setRoot('tests')
 
         // Move the build types to build-types/<type>
         // For instance, build-types/debug/java, build-types/debug/AndroidManifest.xml, ...
