@@ -2,7 +2,7 @@ package com.xpple.graduates.ui;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
+import androidx.fragment.app.FragmentTransaction;
 
 import com.ktplay.open.KTPlay;
 import com.xpple.graduates.CustomApplication;
