@@ -2,7 +2,7 @@ package com.xpple.graduates.view;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+import androidx.fragment.app.FragmentActivity;
 import android.util.DisplayMetrics;
 
 import com.xpple.graduates.CustomApplication;
